Working model with MLP Classifier

+++ app.py
@@ -17,7 +17,7 @@ def process(form_data):
         if index in continuous_column:
             if len(value)==0:
                 value = 0
-            data[index] = int(value)
+            data[index] = float(value)
     df = pd.DataFrame(data=[data.values],columns=data.index)
 
     return df
@@ -27,7 +27,6 @@ def predict():
   
     df = process(request.form)
     prediction = model.predict_proba(df)[0]
-    # prediction= 86.5
     return render_template('predict.html', prediction = round(prediction[1]*100,2))
